24 FastAPI/theme/project/requirements.txt
fastapi
uvicorn
requests
aiologger
aiofiles
python-dotenv
pydantic-settings
sqlalchemy
aiosqlite
alembic
passlib
PyJWT
python-multipart
httpx